pyproject.toml
[project]
name = "langgraph-customer-support"
version = "0.1.0"
description = "Zero to Hero LangGraph Tutorial - Customer Support Agent System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.45",
    "langchain>=0.3.7",
    "langchain-google-genai>=2.0.5",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []


